split mainnet/onyx/altona/e2e configs into different files
