+++ shared/params/BUILD.bazel
@@ -7,7 +7,13 @@ go_library(
         "config.go",
         "io_config.go",
         "loader.go",
+        "mainnet_config.go",
+        "minimal_config.go",
         "network_config.go",
+        "testnet_altona_config.go",
+        "testnet_e2e_config.go",
+        "testnet_onyx_config.go",
+        "testutils.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/shared/params",
     visibility = ["//visibility:public"],
