+++ shared/params/testnet_onyx_config.go
@@ -1,0 +1,20 @@
+package params
+
+// UseOnyxNetworkConfig uses the Onyx specific network config.
+func UseOnyxNetworkConfig() {
+	cfg := BeaconNetworkConfig().Copy()
+	cfg.ContractDeploymentBlock = 2844925
+	cfg.DepositContractAddress = "0x0F0F0fc0530007361933EaB5DB97d09aCDD6C1c8"
+	cfg.BootstrapNodes = []string{"enr:-Ku4QMKVC_MowDsmEa20d5uGjrChI0h8_KsKXDmgVQbIbngZV0idV6_RL7fEtZGo-kTNZ5o7_EJI_vCPJ6scrhwX0Z4Bh2F0dG5ldHOIAAAAAAAAAACEZXRoMpD1pf1CAAAAAP__________gmlkgnY0gmlwhBLf22SJc2VjcDI1NmsxoQJxCnE6v_x2ekgY_uoE1rtwzvGy40mq9eD66XfHPBWgIIN1ZHCCD6A"}
+	OverrideBeaconNetworkConfig(cfg)
+}
+
+// OnyxConfig returns the configuration to be used in the main network.
+func OnyxConfig() *BeaconChainConfig {
+	return mainnetBeaconConfig
+}
+
+// UseMainnetConfig for beacon chain services.
+func UseOnyxConfig() {
+	beaconConfig = MainnetConfig().Copy()
+}
